feat(Tracking): add device details and dominant controller tracking
The new abstract DeviceDetailsRecord sets a framework for holding
common information about a device such as model, connected status
and battery status.

The LinkedAliasAssociationCollection now can hold a DeviceDetailsRecord
for the headset, left and right controller. CameraRig packages
will provide a concrete implementation of the DeviceDetailsRecord for
the headset and controllers of that CameraRig supported SDK and then
linked to the CameraRig prefab.

This will mean that the CameraRig prefab will then be able to know
about the details of the headset and controllers.

A new DominantControllerObserver component has been added as well
that holds a left/right controller DeviceDetailsRecord as the soruce
and will be able to determine whether the left or right controller is
the current dominant controller. This component is also processable
so can be run on a Moment to check periodically if the dominant
controller has changed.