feat(Tracking): store dominant controller on linked alias collection
A DominantControllerObserver can be stored on a
LinkedAliasAssociationCollection to make it easier to access the
current component that is checking for the dominant controller.

The debug logs have also been removed from the XRDeviceProperties that
warned about methods not being available as they could just end up
spamming the console and its not completely necessary.