fix(Cast): prevent casts not resetting on disable
The LineCasts could get into a situation where they would not work
after the component was disabled. This is due to the RayCast target
not being reset to null when the component is disabled.

Also, extra checks to ensure the points arrays are actually set to
the correct values are now done to prevent any exceptions that may
occur.