fix(Tracking): add time limit to rigidbody velocity tests
The RigidbodyVelocityTest can occasionally fail in the diverge tests as
it's relying on the physics engine to work perfectly under test
conditions and it causes the test to hang indefinitely. This has been
fixed by simply adding a counter that will cause the while loop to end
if the timer exceeds the limit.