fix(Data): allow list drawer to save changes
The Observable List Drawer has an issue in Unity 2021.2 where making
changes to a linked list will not save due to the way the serialized
object is obtained as new every frame, this means that the object
can be out of date and load in the old changes.

This also was a cause of an issue where having the linked list
component on the same component as the actual list meant it would not
save changes.

The fix is to only get the new version of the serialized list object
if any changes have been made and therefore it won't attempt to get
it every frame.