feat(Malimbe): remove property serialization and xml documentation
The Malimbe PropertySerializationAttribute and XMLDocumentation
attribute have been removed as a dependency and the property
serialization is now done manually by providing a private backing
field that is used within the related property in the same naming
convention that Malimbe used so all references should be remembered.

XML Documentation is now simply duplicated into the Unity `Toolkit`
attribute.

This commit may have some unwanted outcomes as the MemberChange
attributes still exist and may not weave correctly with the hard coded
property setters.