feat(Malimbe): remove MemberChangeMethod attribute dependency
The Malimbe MemberChangeMethod attribute has been removed as a
dependency and instead every property now manually calls the
`OnBefore` before the value is set and the `OnAfter` after the value
is set.

These are also wrapped in a check to ensure the application is playing
and the component is active and enabled to copy the behaviour of how
the Malimbe attribute worked.