feat(Malimbe): remove Malimbe custom inspector
BREAKING CHANGE: This removes the last remaining elements of
Malimbe and whilst it does not cause any breaking changes within
Zinnia, it removes Malimbe as a dependency which other projects that
rely on Zinnia may piggy back off this Malimbe dependency so it will
break any project like that.

The Malimbe custom inspector has been replaced with the new
ZinniaInspector which performs pretty much the same task except
instead of looking for the MemberChange attributes it now specifically
expects `OnBefore` and `OnAfter` property methods to be explicityly
named with the property name followed by the suffix of `Change`.

E.g.

```
public bool MyProperty {};

protected virtual void OnBeforeMyPropertyChange();
```

The dependency of Malimbe has also been removed as it is no longer
used anywhere now.