fix(ZinniaInspector): catch unwanted exceptions from Unity
There are a couple of oddities where Unity will thrown an exception
using the custom editor if a Horizontal or Vertical group is started
and then another Unity inspector element is opened it seems not to
close the previous opened group causing a mismatch somewhere.

This isn't technically a fix for the problem but just a way of removing
the error for now.