feat(Data): set custom true and false values for boolean to float
The BooleanToFloat component now has the ability to set the false and
true float value rather than it just being hard coded to 0f for false
and 1f for true.

The FloatToBoolean PositiveBounds property has also been made public
as it should be so it can be set via code if required.