feat(Utility): add ability for countdown timer to start on enabled
Previously, the CountdownTimer component would rely on the
BehaviourEnabledObserver component to let it start the timer when
the component became enabled. But the use case for wanting a timer
to start when the component is enabled is so common that it may
as well simply be an option on the CountdownTimer component itself.

This makes it easier to simply start a timer, but it can also be
disabled so the use of the BehaviourEnabledObserver can still be
used if required. It is also set to false by default so will not
cause any breaking changes.