feat(Action): add sources contains method
The Action now has a SourcesContains method that allows a generic
action to be checked if its sources contain a given action.

This is on the base class so can be used regardless of any action
typing.

The RemoveSource method also now returns a bool based on the results
of the internal `List.Remove` call. This is also now used in the
ActionRegistrar to determine if the remove/unregister events should
be raised as if no source is removed then no event should be raised.