fix(Tests): ensure all quaternion and vectors use comparators
All of the tests that compare vectors or quaternions have now been
updated so they no longer use ToString to compare in the
Assert.AreEqual as this is not robust enough to always work. Instead
they all now use equality comparators.

All of the gameobjects created by the tests now also are named so it
is easier to see which test is running and where any cleanup issues
may be occurring.

The `UnityEngine.Assertions.Assert` has also been removed from the
tests as it is not required.

Any test that destroys the `subject` object on teardown has also been
updated to not do this as the subject gets destroyed automatically
when the `containingObject` GameObject gets destroyed.