feat(Tracking): add pivot rotation mirror to direction modifier
The DirectionModifier component now has a new property of
PivotRotationMirror which takes a GameObject that will mirror the
rotation of the given Pivot GameObject.

This is useful if for instance the follow logic is running on a
precision point basis and that precision point object needs to
replicate the rotation given by the Pivot.