fix(Tracking): stop artificial velocity applier running when disabled
The ArtificialVelocityApplier was still applying velocity via the
`Apply` method even if the component was disabled.

This is because originally it used the `[RequiresBehaviourState]`
Malimbe tag but when it was separated out from the process parent,
this was not updated when malimbe was removed.

This has been fixed now by ensuring the valid state of the object
is present before running the `Apply` method.