feat(Extension): provide alternative method for isActiveAndEnabled
It would seem that the behaviour.isActiveAndEnabled property does not
always return what is expected. If a GameObject is active and the
behaviour is enabled then isActiveAndEnabled can still return false
whilst in the OnEnable state.

It would seem that isActiveAndEnabled does not return true until
all of the relevant initial Unity moments have passed for an Object.

To counter this, a new BehaviourExtension has been added called
`CheckIsActiveAndEnabled()` that will do an absolute check on
the GameObject active state and the behaviour enabled state.

If the old isActiveAndEnabled check is still required then it can
be switched to by a scripting define symbol of:

`ZINNIA_USE_ISACTIVEANDENABLED`

All of the relevant files that use `isActiveAndEnabled` have now been
updated to use the new `CheckIsActiveAndEnabled()` extension method.