feat(Utility): add pipeline material applier component
The PipelineMaterialApplier will set the Materials on a given Renderer
based on the current render pipeline being used. It gets the materials
to apply from a given list of materials that has a string group name
that matches the desired render pipeline name.

If no render pipeline is provided then the matched name will be:

* `default`

There is also an editor menu item that will attempt to find all
PipelineMaterialApplier components in the scene and run them to
ensure all materials on the renderers are updated to match the current
render pipeline.