fix(Utility): limit pipeline material applier to unity 2020 or above
The mechanism within the ScenePipelineMaterialApplier uses the
`FindObjectsOfType` method that in Unity 2019.2 had the additional
parameter of `findInactive` added, which is required for this
functionality to work.

However, in Unity 2019.4 this parameter is randomly removed, probably
due to poor quality control from Unity, so this means that this
functionality works in 2019.2, 2019.3 but throws an error in 2019.4.

The feature is reintroduced in 2020.1, so the fix is just to support
this functionality in 2020.1 as Unity won't back port a fix.