feat(Cast): add cursor lock and transition duration to points cast
The PointsCast now has a cursor lock function and a transition duration
which allows the cursor to be locked to a location until the pointer
moves a certain threshold distance away when the cursor location will
then be updated, this can be used to prevent jittering or keep a solid
lock on an object. The transition duration allows the pointer
destination cursor to transition over time to the new destination
providing a smoothing feature for the pointer cursor.

The FixedLineCast has now been deprecated as the StraightLineCast now
handles the concept of being a fixed line as it's basically the same
thing and makes it more complex to implement functionality that is
shared between the two concepts.

The StraightLineCast now has the ability to fix the line at a set
length and optionally ignore any new targets once fixed.

Along with the cursor lock and transition duration, it's also possible
to apply a drag effect to the straight line where the straight line
actually bends to try and keep the straight line from the origin
to the offset destination.