feat(Mutation): emit pre/post mutation events on property mutators
The TransformPropertyMutator now emits an event on pre mutation with
the value that it is going to mutate the target to. It also emits an
event when the mutation has occurred with this new value.

It is also possible to prevent a mutation from occuring by setting the
new Allow Mutation boolean property, which will still do the pre
mutation event but will then skip the actual mutation of the target
and not raise the post mutation event, but instead raise a mutation
skipped event.

This makes it possible to check to see if a property mutation will
provide an undesired result in the pre mutation and then prevent
the mutation from occuring by turning off the Allow Mutation.

Some refactoring of the internal calls has been done too so it is
cleaner and less code repetition.