feat(Tracking): apply distance offset on position difference rotation
The TransformPositionDifferenceRotation logic now has a source to
offset maximum distance setting that will be activated if the
`Apply Offset` is to true. This will ensure the rotation will only
be applied when the source and offset are within the given distances
and anything outside of those distances will be ignored.