fix(Data): prevent null exception in collapsible drawer in 2022.3
Due to a change in Unity 2022.3, there is now no actual need to do
the reflection for replacing a default drawer and in fact the
attempt to get `DrawerKeySet` will return a null, thus causing the
existing error.

The fix is to simply do a null check to prevent the error and the
reflection will simply be ignored on versions of Unity of 2022.3
and above.